ingestion/requirements.txt
psycopg2-binary==2.9.9
pymongo==4.6.1
requests==2.31.0
minio==7.2.5
pandas==2.2.0
python-dateutil==2.8.2
